requirements.txt
graphviz
joblib==1.1.0
librosa
matplotlib
numpy
opencv_python
PyYAML==6.0
# scikit_learn==1.2.2
scipy
SoundFile
tensorboardX
# tensorflow==2.6.1
# torch==1.10.1
# torchvision
tqdm
